+++ core/new-runner/src/main/java/com/cognifide/aet/runner/RunnerConfiguration.java
@@ -13,7 +13,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.runner.configs;
+package com.cognifide.aet.runner;
 
 import java.util.Map;
 import org.apache.felix.scr.annotations.Activate;
