+++ core/new-runner/src/test/java/com/cognifide/aet/runner/processing/steps/StepManagerTest.java
@@ -24,7 +24,7 @@ import static org.mockito.internal.verification.VerificationModeFactory.times;
 import com.cognifide.aet.communication.api.ProcessingError;
 import com.cognifide.aet.communication.api.metadata.Suite;
 import com.cognifide.aet.communication.api.queues.JmsConnection;
-import com.cognifide.aet.runner.configs.RunnerConfiguration;
+import com.cognifide.aet.runner.RunnerConfiguration;
 import com.cognifide.aet.runner.processing.TimeoutWatch;
 import com.cognifide.aet.runner.processing.data.SuiteIndexWrapper;
 import com.cognifide.aet.runner.scheduler.CollectorJobSchedulerService;
